sndh: Fix timer to frequency cast for C++ compilers
include/psgplay/sndh.h: In function ‘sndh_timer u32_to_sndh_timer(uint32_t)’:
include/psgplay/sndh.h:328:36: warning: narrowing conversion of ‘(((unsigned int)timer) >> 8)’ from ‘unsigned int’ to ‘int’ [-Wnarrowing]
  328 |                 .frequency = timer >> 8,
      |                              ~~~~~~^~~~

+++ include/psgplay/sndh.h
@@ -325,7 +325,7 @@ static inline struct sndh_timer u32_to_sndh_timer(const uint32_t timer)
 {
 	return (struct sndh_timer) {
 		.type = (enum sndh_timer_type)(timer & 0xff),
-		.frequency = timer >> 8,
+		.frequency = (int)(timer >> 8),
 	};
 }
